+++ tabulas-ext/src/main/scala/de/tudresden/inf/lat/tabulas/ext/main/Main.scala
@@ -18,6 +18,7 @@ object Main {
       RxYamlExtension(),
       YamlExtension(),
       GuessExtension(),
+      JsonSchemaExtension(),
       JsonExtension(),
       PropertiesExtension(),
       OldFormatExtension(),